:recycle: refactor(search): Extract `usePatientsSearchResults`.
